% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{assign.list}
\alias{assign.list}
\title{FUNCTION_TITLE}
\usage{
assign.list(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
