% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoNameClusters.R
\name{AutoNameClusters}
\alias{AutoNameClusters}
\title{FUNCTION_TITLE}
\usage{
AutoNameClusters(
  Seu,
  auto.cluster = "EC",
  species = "Mouse",
  name.orig = "seurat_clusters",
  name.new = "auto_cluster"
)
}
\arguments{
\item{Seu}{PARAM_DESCRIPTION}

\item{auto.cluster}{PARAM_DESCRIPTION, Default: 'EC'}

\item{species}{PARAM_DESCRIPTION, Default: 'Mouse'}

\item{name.orig}{PARAM_DESCRIPTION, Default: 'seurat_clusters'}

\item{name.new}{PARAM_DESCRIPTION, Default: 'auto_cluster'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
