% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistrBar.R
\name{BarOfCluster}
\alias{BarOfCluster}
\title{FUNCTION_TITLE}
\usage{
BarOfCluster(meta, cluster, origin, method = "value", do.heatmap = T)
}
\arguments{
\item{meta}{PARAM_DESCRIPTION}

\item{cluster}{PARAM_DESCRIPTION}

\item{origin}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'value'}

\item{do.heatmap}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
