% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_basic.R
\name{BuildAUCRank}
\alias{BuildAUCRank}
\title{Build AUCell Rank}
\usage{
BuildAUCRank(seu, slot = "counts", assay = "RNA")
}
\arguments{
\item{seu}{Seurat object}

\item{slot}{Slot to pull feature data for, Default: 'counts'}

\item{assay}{Name of assay to use, Default: 'RNA'}
}
\value{
Seurat object.
}
\description{
Build AUCell ranking matrix using 'AUCell' package: \n
Aibar et al. (2017) SCENIC: single-cell regulatory network inference and clustering.
Nature Methods. doi: 10.1038/nmeth.4463 \n
Aibar. et al. (2016) AUCell: Analysis of 'gene set' activity in single-cell RNA-seq data.
R/Bioconductor package.
}
\details{
AUCell ranking matrix is saved in seu@misc$AUCell[["cells_rankings"]]
}
\examples{
pbmc <- BuildAUCRank(pbmc)
pbmc@misc$AUCell[["cells_rankings"]]
}
