% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_basic.R
\name{calcAUC_matrix}
\alias{calcAUC_matrix}
\title{FUNCTION_TITLE}
\usage{
calcAUC_matrix(
  GenesetList,
  rankings,
  nCores = 1,
  verbose = TRUE,
  n.items.part = NULL
)
}
\arguments{
\item{GenesetList}{PARAM_DESCRIPTION}

\item{rankings}{PARAM_DESCRIPTION}

\item{nCores}{PARAM_DESCRIPTION, Default: 1}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{n.items.part}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}
 \code{\link[rlist]{list.rbind}}
}
