% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcStats.R
\name{CalcScoreGeneral}
\alias{CalcScoreGeneral}
\alias{CalcScoreGeneral_v3}
\title{CalcScoreGeneral}
\usage{
CalcScoreGeneral(matr, f, method = "zscore", exp.transform = F, t = F)

CalcScoreGeneral_v3(
  seu,
  features,
  group.by = NULL,
  cells = NULL,
  slot = "data",
  method = "zscore",
  exp.transform = F
)
}
\description{
Alias of \code{\link[SeuratExtend:CalcStats]{CalcStats()}}
}
\seealso{
\code{\link[SeuratExtend:CalcStats]{CalcStats()}}
}
