% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcScoreGeneral.R
\name{CalcScoreGeneral_v3}
\alias{CalcScoreGeneral_v3}
\title{FUNCTION_TITLE}
\usage{
CalcScoreGeneral_v3(
  Seu,
  features,
  group.by,
  method,
  assay = "RNA",
  exp.transform = F
)
}
\arguments{
\item{Seu}{PARAM_DESCRIPTION}

\item{features}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION, Default: 'RNA'}

\item{exp.transform}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
