% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{CellphoneDB_GenerateCustomDB}
\alias{CellphoneDB_GenerateCustomDB}
\title{FUNCTION_TITLE}
\usage{
CellphoneDB_GenerateCustomDB(Reactome_interactions_filtered, path = getwd())
}
\arguments{
\item{Reactome_interactions_filtered}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: getwd()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
