% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{CellphoneDB_Plots}
\alias{CellphoneDB_Plots}
\title{FUNCTION_TITLE}
\usage{
CellphoneDB_Plots(seu, sender, receiver, group.by, top_n = 20)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{sender}{PARAM_DESCRIPTION}

\item{receiver}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{top_n}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
