% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{CellphoneDB_Plots_Circlize}
\alias{CellphoneDB_Plots_Circlize}
\title{FUNCTION_TITLE}
\usage{
CellphoneDB_Plots_Circlize(
  seu = NULL,
  output = NULL,
  sender = NULL,
  receiver = NULL,
  nLink = 50,
  ignore = NULL
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION, Default: NULL}

\item{sender}{PARAM_DESCRIPTION, Default: NULL}

\item{receiver}{PARAM_DESCRIPTION, Default: NULL}

\item{nLink}{PARAM_DESCRIPTION, Default: 50}

\item{ignore}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
