% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotToGenesymbol.R
\name{check_sub_loc}
\alias{check_sub_loc}
\title{FUNCTION_TITLE}
\usage{
check_sub_loc(
  Symbol,
  spe = getOption("spe"),
  loc = c("Secreted", "Membrane", "Cell membrane")
)
}
\arguments{
\item{Symbol}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}

\item{loc}{PARAM_DESCRIPTION, Default: c("Secreted", "Membrane", "Cell membrane")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
