% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\name{color_iwh}
\alias{color_iwh}
\title{color_iwh}
\usage{
color_iwh(n, col.space = c("default", "intense", "pastel"), set = 1)
}
\arguments{
\item{n}{How many colors to generate}

\item{col.space}{Color space, Default: c("default", "intense", "pastel")}

\item{set}{Several random presets, Default: 1}
}
\value{
Vector of colors
}
\description{
Generate color presets from 'I want hue' online tool
}
\details{
Random color presets generated from: https://medialab.github.io/iwanthue/
}
\examples{
color_iwh(10)
}
