% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\name{color_pro}
\alias{color_pro}
\title{Professional discrete color presets by 'I want hue'}
\usage{
color_pro(
  n,
  col.space = c("default", "light", "red", "yellow", "green", "blue", "purple"),
  sort = c("hue", "diff"),
  set = 1
)
}
\arguments{
\item{n}{How many colors to generate}

\item{col.space}{Color space, Options: "default", "light", "red", "yellow", "green", "blue", or "purple"}

\item{sort}{Sort colors by hue (default) or differentiation, Options: "hue", "diff"}

\item{set}{Several random presets, Default: 1}
}
\value{
Vector of colors
}
\description{
Generate professional color presets from 'I want hue' color tool
}
\details{
Random color presets generated from: https://medialab.github.io/iwanthue/
}
\examples{
color_pro(10)
}
