% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Python.R
\name{create_condaenv_seuratextend}
\alias{create_condaenv_seuratextend}
\title{Create the 'seuratextend' Conda Environment}
\usage{
create_condaenv_seuratextend()
}
\value{
Does not return any value; it creates and configures a new Conda environment.
}
\description{
Initializes a Conda environment named 'seuratextend' which includes all necessary Python packages for running `scVelo`, `Palantir`, and `CellRank` via `SeuratExtend`. This function is ideal for users setting up their analytical environment for the first time or those who need to ensure they have a compatible environment setup.
}
\details{
When you run functions from `SeuratExtend` that require Python integrations, such as `scVelo`, `Palantir`, or `CellRank`, you might need a specific set of Python packages installed. This function facilitates the setup of a Conda environment specifically tailored for these tools, managed through `reticulate`. The environment setup is automatic and checks your operating system to configure appropriately. It is supported on Windows, macOS, and Linux (Ubuntu 20.04), ensuring broad compatibility and ease of setup.
}
\examples{
\dontrun{
# To manually create the 'seuratextend' Conda environment, execute:
create_condaenv_seuratextend()
}
}
