% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimplot3d.R
\name{Dimplot3d}
\alias{Dimplot3d}
\title{FUNCTION_TITLE}
\usage{
Dimplot3d(
  Seu,
  group.by = "seurat_clusters",
  assay = NULL,
  reduction = "pca",
  dims = NULL,
  text_shift = 0.1,
  size = 16,
  width = 640,
  legend3d = T,
  writeWebGL = F,
  filename = NULL,
  slingPseudotime = NULL
)
}
\arguments{
\item{Seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION, Default: 'seurat_clusters'}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}

\item{reduction}{PARAM_DESCRIPTION, Default: 'pca'}

\item{dims}{PARAM_DESCRIPTION, Default: NULL}

\item{text_shift}{PARAM_DESCRIPTION, Default: 0.1}

\item{size}{PARAM_DESCRIPTION, Default: 16}

\item{width}{PARAM_DESCRIPTION, Default: 640}

\item{legend3d}{PARAM_DESCRIPTION, Default: T}

\item{writeWebGL}{PARAM_DESCRIPTION, Default: F}

\item{filename}{PARAM_DESCRIPTION, Default: NULL}

\item{slingPseudotime}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
