% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblToGenesymbol.R
\name{EnsemblToGenesymbol}
\alias{EnsemblToGenesymbol}
\title{FUNCTION_TITLE}
\usage{
EnsemblToGenesymbol(Ensembl, spe = getOption("spe"), mirror = NULL)
}
\arguments{
\item{Ensembl}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}

\item{mirror}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
