% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePlot.R
\name{FeaturePlot3.grid}
\alias{FeaturePlot3.grid}
\title{FUNCTION_TITLE}
\usage{
FeaturePlot3.grid(
  seu,
  features,
  color = c("ryb", "rgb"),
  color.range = c(0.1, 0.95),
  reduction = "umap",
  order = T,
  pt.size = 0.1
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{features}{PARAM_DESCRIPTION}

\item{color}{PARAM_DESCRIPTION, Default: c("ryb", "rgb")}

\item{color.range}{PARAM_DESCRIPTION, Default: c(0.1, 0.95)}

\item{reduction}{PARAM_DESCRIPTION, Default: 'umap'}

\item{order}{PARAM_DESCRIPTION, Default: T}

\item{pt.size}{PARAM_DESCRIPTION, Default: 0.1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
