% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePlot.R
\name{FeaturePlot3}
\alias{FeaturePlot3}
\title{FUNCTION_TITLE}
\usage{
FeaturePlot3(
  seu,
  feature.1 = NA,
  feature.2 = NA,
  feature.3 = NA,
  color = c("ryb", "rgb"),
  color.range = c(0.1, 0.9),
  reduction = "umap",
  order = T
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{feature.1}{PARAM_DESCRIPTION, Default: NA}

\item{feature.2}{PARAM_DESCRIPTION, Default: NA}

\item{feature.3}{PARAM_DESCRIPTION, Default: NA}

\item{color}{PARAM_DESCRIPTION, Default: c("ryb", "rgb")}

\item{color.range}{PARAM_DESCRIPTION, Default: c(0.1, 0.9)}

\item{reduction}{PARAM_DESCRIPTION, Default: 'umap'}

\item{order}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
