% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gene_expressed}
\alias{gene_expressed}
\title{FUNCTION_TITLE}
\usage{
gene_expressed(
  ident,
  seu,
  feature = rownames(seu),
  group.by = NULL,
  pct = 0.1,
  DefaultAssay = "RNA"
)
}
\arguments{
\item{ident}{PARAM_DESCRIPTION}

\item{seu}{PARAM_DESCRIPTION}

\item{feature}{PARAM_DESCRIPTION, Default: rownames(seu)}

\item{group.by}{PARAM_DESCRIPTION, Default: NULL}

\item{pct}{PARAM_DESCRIPTION, Default: 0.1}

\item{DefaultAssay}{PARAM_DESCRIPTION, Default: 'RNA'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
