% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gene_percent}
\alias{gene_percent}
\title{FUNCTION_TITLE}
\usage{
gene_percent(seu, feature, group.by, cluster)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{feature}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{cluster}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
