% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysisReactome.R
\name{GeneSetAnalysisReactome}
\alias{GeneSetAnalysisReactome}
\title{FUNCTION_TITLE}
\usage{
GeneSetAnalysisReactome(
  seu = NULL,
  parent = "All",
  spe = getOption("spe"),
  ratio = 0.4,
  n.min = 1,
  n.max = Inf,
  only.end.terms = F,
  slot = "counts",
  assay = "RNA",
  nCores = getOption("nCores"),
  export_to_matrix = F,
  verbose = TRUE,
  n.items.part = 5e+05/ncol(seu) * parallel::detectCores()
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION, Default: NULL}

\item{parent}{PARAM_DESCRIPTION, Default: 'All'}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}

\item{ratio}{PARAM_DESCRIPTION, Default: 0.4}

\item{n.min}{PARAM_DESCRIPTION, Default: 1}

\item{n.max}{PARAM_DESCRIPTION, Default: Inf}

\item{only.end.terms}{PARAM_DESCRIPTION, Default: F}

\item{slot}{PARAM_DESCRIPTION, Default: 'counts'}

\item{assay}{PARAM_DESCRIPTION, Default: 'RNA'}

\item{nCores}{PARAM_DESCRIPTION, Default: getOption("nCores")}

\item{export_to_matrix}{PARAM_DESCRIPTION, Default: F}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{n.items.part}{PARAM_DESCRIPTION, Default: 5e+05/ncol(seu) * parallel::detectCores()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}
}
