% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEndTermsReactome.R
\name{GetEndTermsReactome}
\alias{GetEndTermsReactome}
\title{FUNCTION_TITLE}
\usage{
GetEndTermsReactome(term, spe = getOption("spe"))
}
\arguments{
\item{term}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
