% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nichenetr.R
\name{GetSeuratDotplotData}
\alias{GetSeuratDotplotData}
\title{FUNCTION_TITLE}
\usage{
GetSeuratDotplotData(
  object,
  assay = NULL,
  features,
  cols = "RdYlBu",
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  group.by = NULL,
  split.by = NULL,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION}

\item{cols}{PARAM_DESCRIPTION, Default: 'RdYlBu'}

\item{col.min}{PARAM_DESCRIPTION, Default: -2.5}

\item{col.max}{PARAM_DESCRIPTION, Default: 2.5}

\item{dot.min}{PARAM_DESCRIPTION, Default: 0}

\item{dot.scale}{PARAM_DESCRIPTION, Default: 6}

\item{group.by}{PARAM_DESCRIPTION, Default: NULL}

\item{split.by}{PARAM_DESCRIPTION, Default: NULL}

\item{scale.by}{PARAM_DESCRIPTION, Default: 'radius'}

\item{scale.min}{PARAM_DESCRIPTION, Default: NA}

\item{scale.max}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
