% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVA.R
\name{GSEAplot}
\alias{GSEAplot}
\title{FUNCTION_TITLE}
\usage{
GSEAplot(
  seu,
  group.by,
  geneset,
  ident.1 = NULL,
  ident.2 = NULL,
  logFC.cutoff = c(-1, 1),
  title = deparse(substitute(geneset))
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{geneset}{PARAM_DESCRIPTION}

\item{ident.1}{PARAM_DESCRIPTION, Default: NULL}

\item{ident.2}{PARAM_DESCRIPTION, Default: NULL}

\item{logFC.cutoff}{PARAM_DESCRIPTION, Default: c(-1, 1)}

\item{title}{PARAM_DESCRIPTION, Default: deparse(substitute(geneset))}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
