% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesymbolConversion.R
\name{HumanToMouseGenesymbol}
\alias{HumanToMouseGenesymbol}
\title{FUNCTION_TITLE}
\usage{
HumanToMouseGenesymbol(human_genes, mirror = NULL, local.mode = T)
}
\arguments{
\item{human_genes}{PARAM_DESCRIPTION}

\item{mirror}{PARAM_DESCRIPTION, Default: NULL}

\item{local.mode}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
