% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{ImportCellphoneDBOutputs}
\alias{ImportCellphoneDBOutputs}
\title{FUNCTION_TITLE}
\usage{
ImportCellphoneDBOutputs(seu = NULL, dir = "out")
}
\arguments{
\item{seu}{PARAM_DESCRIPTION, Default: NULL}

\item{dir}{PARAM_DESCRIPTION, Default: 'out'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
