% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanToMouseGenesymbol.R
\name{MouseToHumanGenesymbol}
\alias{MouseToHumanGenesymbol}
\title{FUNCTION_TITLE}
\usage{
MouseToHumanGenesymbol(mouse_genes, mirror = NULL)
}
\arguments{
\item{mouse_genes}{PARAM_DESCRIPTION}

\item{mirror}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
