% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nichenetr.R
\name{Nichenetr_Plots}
\alias{Nichenetr_Plots}
\title{FUNCTION_TITLE}
\usage{
Nichenetr_Plots(nichenet_results, seu.sender = NULL, ident.rcv = NULL)
}
\arguments{
\item{nichenet_results}{PARAM_DESCRIPTION}

\item{seu.sender}{PARAM_DESCRIPTION, Default: NULL}

\item{ident.rcv}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cowplot]{plot_grid}}
 \code{\link[ggpubr]{as_ggplot}},\code{\link[ggpubr]{get_legend}}
}
