% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pair_rev}
\alias{pair_rev}
\title{FUNCTION_TITLE}
\usage{
pair_rev(pairs, sep = "_")
}
\arguments{
\item{pairs}{PARAM_DESCRIPTION}

\item{sep}{PARAM_DESCRIPTION, Default: '_'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
