% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysisGO.R,
%   R/GeneSetAnalysisReactome.R
\name{RenameGO}
\alias{RenameGO}
\alias{RenameReactome}
\title{Convert GO/Reactome Pathway IDs to Full Names}
\usage{
RenameGO(term, add_id = T, add_n_gene = T, spe = getOption("spe"))

RenameReactome(term, add_id = T, add_n_gene = T, spe = getOption("spe"))
}
\arguments{
\item{term}{GO/Reactome IDs, which can be either a character vector or a matrix (with IDs in rownames).}

\item{add_id}{Logical: Should the IDs be displayed? Default: TRUE.}

\item{add_n_gene}{Logical: Should the number of genes in each geneset be displayed? Default: TRUE.}

\item{spe}{Species, either 'human' or 'mouse'. Default: getOption("spe").}
}
\value{
Returns either a character vector or a matrix, depending on the input.
}
\description{
Converts GO or Reactome IDs to their respective full geneset or pathway names.
}
\details{
For more examples, see \code{\link[SeuratExtend:GeneSetAnalysis]{GeneSetAnalysis()}}.
}
\examples{
library(SeuratExtend)

RenameGO(c("GO:0002376","GO:0050896"), spe = "human")

RenameReactome(c("R-HSA-109582","R-HSA-112316"), spe = "human")
}
