% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameReactome.R
\name{RenameReactome}
\alias{RenameReactome}
\title{FUNCTION_TITLE}
\usage{
RenameReactome(term, spe = getOption("spe"))
}
\arguments{
\item{term}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
