% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunBasicUmap.R
\name{RunBasicUmap}
\alias{RunBasicUmap}
\title{FUNCTION_TITLE}
\usage{
RunBasicUmap(
  path = getwd(),
  project,
  sub.path = "/outs/filtered_feature_bc_matrix",
  saveRDS = T,
  save.path = "rds",
  spe = getOption("spe"),
  min.cells = 1,
  min.features = 200,
  dims = c(1:20),
  resolution = 0.5
)
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: getwd()}

\item{project}{PARAM_DESCRIPTION}

\item{sub.path}{PARAM_DESCRIPTION, Default: '/outs/filtered_feature_bc_matrix'}

\item{saveRDS}{PARAM_DESCRIPTION, Default: T}

\item{save.path}{PARAM_DESCRIPTION, Default: 'rds'}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}

\item{min.cells}{PARAM_DESCRIPTION, Default: 1}

\item{min.features}{PARAM_DESCRIPTION, Default: 200}

\item{dims}{PARAM_DESCRIPTION, Default: c(1:20)}

\item{resolution}{PARAM_DESCRIPTION, Default: 0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
