% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{RunCellphoneDB}
\alias{RunCellphoneDB}
\title{FUNCTION_TITLE}
\usage{
RunCellphoneDB(seu, group.by, database = "cellphonedb_mouse_cytokines")
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{database}{PARAM_DESCRIPTION, Default: 'cellphonedb_mouse_cytokines'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
