% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiffusionMap.R
\name{RunDiffusionMap}
\alias{RunDiffusionMap}
\title{FUNCTION_TITLE}
\usage{
RunDiffusionMap(Seu, assay = NULL)
}
\arguments{
\item{Seu}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
