% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVA.R
\name{RunGSVA}
\alias{RunGSVA}
\title{FUNCTION_TITLE}
\usage{
RunGSVA(matr, genesets, min.sz = 0, max.sz = Inf)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{genesets}{PARAM_DESCRIPTION}

\item{min.sz}{PARAM_DESCRIPTION, Default: 0}

\item{max.sz}{PARAM_DESCRIPTION, Default: Inf}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
