% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVA.R
\name{RunGSVA_v3}
\alias{RunGSVA_v3}
\title{FUNCTION_TITLE}
\usage{
RunGSVA_v3(
  seu,
  group.by,
  genesets,
  min.sz = 0,
  max.sz = Inf,
  title = "genesets"
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{genesets}{PARAM_DESCRIPTION}

\item{min.sz}{PARAM_DESCRIPTION, Default: 0}

\item{max.sz}{PARAM_DESCRIPTION, Default: Inf}

\item{title}{PARAM_DESCRIPTION, Default: 'genesets'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
