% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palantir.R
\name{RunPalantirDiffusionMap}
\alias{RunPalantirDiffusionMap}
\title{(Deprecated) Run Palantir Diffusion Map}
\usage{
RunPalantirDiffusionMap(seu, reduction = "pca", n_components = 20)
}
\arguments{
\item{seu}{A Seurat object.}

\item{reduction}{A character string specifying the dimension reduction to use. Default: 'pca'.}

\item{n_components}{An integer specifying how many dimensions to use as input. Default: 20.}
}
\value{
A Seurat object with added dm, ms, and t-SNE embeddings.
}
\description{
This is a deprecated version of the Palantir Diffusion Map function.
Please refer to the updated function, \code{\link{Palantir.RunDM}}, for the latest features and improvements.
This older version, in addition to calculating the diffusion map (dm) and multiscale_space (ms),
also calculates t-SNE. To use the t-SNE feature, the \code{multicore-tsne} package needs to be installed in the conda environment.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratExtend)
pbmc <- RunPalantirDiffusionMap(pbmc)
DimPlot(pbmc, reduction = "tsne", label = TRUE)
}
}
\seealso{
\code{\link{Palantir.RunDM}} for the updated version of this function.
}
