% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSlingshot.R
\name{RunSlingshot}
\alias{RunSlingshot}
\title{FUNCTION_TITLE}
\usage{
RunSlingshot(Seu, group.by, reducedDim = "DM", assay = NULL, start.clus = NULL)
}
\arguments{
\item{Seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{reducedDim}{PARAM_DESCRIPTION, Default: 'DM'}

\item{assay}{PARAM_DESCRIPTION, Default: NULL}

\item{start.clus}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
