% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreAndOrder.R
\name{ScoreAndOrder}
\alias{ScoreAndOrder}
\title{FUNCTION_TITLE}
\usage{
ScoreAndOrder(
  matr,
  f,
  method = "zscore",
  order = "p",
  n = Inf,
  p.threshold = 0.05
)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'zscore'}

\item{order}{PARAM_DESCRIPTION, Default: 'p'}

\item{n}{PARAM_DESCRIPTION, Default: Inf}

\item{p.threshold}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
