% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVelo.R
\name{scVelo.AddDimReduc}
\alias{scVelo.AddDimReduc}
\title{FUNCTION_TITLE}
\usage{
scVelo.AddDimReduc(adata, seu, dr = NULL)
}
\arguments{
\item{adata}{PARAM_DESCRIPTION}

\item{seu}{PARAM_DESCRIPTION}

\item{dr}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
