% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVelo.R
\name{scVelo.RunBasic}
\alias{scVelo.RunBasic}
\title{FUNCTION_TITLE}
\usage{
scVelo.RunBasic(loom, save.adata = "adata.obj")
}
\arguments{
\item{loom}{PARAM_DESCRIPTION}

\item{save.adata}{PARAM_DESCRIPTION, Default: 'adata.obj'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
