% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVelo.R
\name{scVelo.SeuratToLoom}
\alias{scVelo.SeuratToLoom}
\title{FUNCTION_TITLE}
\usage{
scVelo.SeuratToLoom(
  seu,
  export.name,
  prefix,
  postfix,
  velocyto.loom.path,
  velocyto.loom.filenames
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{export.name}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION}

\item{postfix}{PARAM_DESCRIPTION}

\item{velocyto.loom.path}{PARAM_DESCRIPTION}

\item{velocyto.loom.filenames}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
