% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchDatabase.R
\name{SearchPathways}
\alias{SearchPathways}
\title{Search Pathways in a Given Geneset List}
\usage{
SearchPathways(genesets, item, type = c("gene", "SetName"))
}
\arguments{
\item{genesets}{A list of genesets where the names are pathway names and the contents are genes.}

\item{item}{A vector containing gene names or (parts of) pathway names to search for.}

\item{type}{Specifies the types of search criteria. This can be "gene", "SetName", or both. Default is c("gene", "SetName").}
}
\value{
A subset of the provided geneset list containing the specified items.
}
\description{
This function filters a provided geneset list based on a specified gene name, or pathway name.
}
\details{
For more details and use-cases, see the examples section.
}
\examples{
library(SeuratExtend)
library(dplyr)
options(max.print = 10)

# Check within the hallmark 50 database for pathways that contain the "CD3D" gene
# or have pathway names that include "interferon".
SearchPathways(genesets = hall50$human, item = c("CD3D", "interferon"))
}
