% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin}
\alias{StackedViolin}
\title{FUNCTION_TITLE}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  ncol = 1,
  lab_fill = "group",
  scales = "free_y",
  type = c("violin", "boxplot"),
  outlier.size = 1
)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{f2}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: 1}

\item{lab_fill}{PARAM_DESCRIPTION, Default: 'group'}

\item{scales}{PARAM_DESCRIPTION, Default: 'free_y'}

\item{type}{PARAM_DESCRIPTION, Default: c("violin", "boxplot")}

\item{outlier.size}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
