% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin_v3}
\alias{StackedViolin_v3}
\title{FUNCTION_TITLE}
\usage{
StackedViolin_v3(
  seu,
  features,
  group.by = "seurat_clusters",
  split.by = NULL,
  cell = NULL,
  ...
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{features}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION, Default: 'seurat_clusters'}

\item{split.by}{PARAM_DESCRIPTION, Default: NULL}

\item{cell}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
