% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotToGenesymbol.R
\name{UniprotToGenesymbol}
\alias{UniprotToGenesymbol}
\title{FUNCTION_TITLE}
\usage{
UniprotToGenesymbol(Uniprot, spe = getOption("spe"))
}
\arguments{
\item{Uniprot}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
