% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Vlnplots.R
\name{VlnPlot2}
\alias{VlnPlot2}
\alias{VlnPlot2.Seurat}
\alias{VlnPlot2.default}
\title{Modified violin plot}
\usage{
VlnPlot2(object, ...)

\method{VlnPlot2}{Seurat}(
  seu,
  features,
  group.by = NULL,
  split.by = NULL,
  cell = NULL,
  slot = "data",
  assay = NULL,
  priority = c("expr", "none"),
  ...
)

\method{VlnPlot2}{default}(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  t = F,
  ncol = 1,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  pt = T,
  hide.outlier = F,
  pt.style = c("jitter", "quasirandom"),
  pt.size = 1,
  pt.alpha = 1,
  strip.position = "top"
)
}
\arguments{
\item{object}{An object; either Seurat object or matrix}

\item{...}{Arguments passed to other methods}

\item{seu}{(Seurat version) Seurat object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData), Default: NULL (All features
in matrix)}

\item{group.by}{(Seurat version) A variable name in meta.data to
group the violin plots by}

\item{split.by}{(Seurat version) A variable name in meta.data to
split the violin plots by}

\item{cell}{(Seurat version) Cell names to use, Default: all cells}

\item{slot}{(Seurat version) Slot to pull feature data for}

\item{assay}{(Seurat version) Name of assay to use, defaults to the active assay}

\item{priority}{(Seurat version) If set to "expr", force fetch data from expression matrix
instead of meta.data}

\item{matr}{Matrix or data frame.Row - features; columns - cells}

\item{f}{Factor or vector. Identity of each cell. Should be the
same length of cells}

\item{f2}{Factor or vector. Similar to \code{f}. A variable to split
the violin plots by. Default: NULL}

\item{t}{If your matrix has features in columns and cells in rows,
you should transpose the matrix first. Default: F}

\item{ncol}{Number of columns if multiple plots are displayed, Default: 1}

\item{lab_fill}{Title of figure legend, Default: 'group'}

\item{scales}{scales parameter passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}
, Default: 'free_y'}

\item{violin}{Whether to plot violin plot, Default: T}

\item{box}{Whether to plot box plot, Default: T}

\item{width}{Width of box plot, Default: 0.9}

\item{pt}{Whether to plot points, Default: T}

\item{hide.outlier}{Whether to hide outlier points of boxplot, Default: F}

\item{pt.style}{Position adjustment, Default: c("jitter", "quasirandom")}

\item{pt.size}{Point size, Default: 1}

\item{pt.alpha}{Point transparency, Default: 1}

\item{strip.position}{Were to put the strip ("top" (default), "bottom", "left",
or "right"). Only use when \code{f2 = NULL}}
}
\value{
ggplot object
}
\description{
Modified violin plots of Seurat style. More compact for multiple variables.
Input can be Seurat object or data frame.
}
\details{
See above
}
\examples{
VlnPlot2(pbmc, features = c("CD3D","CD79A"))

VlnPlot2(pbmc, features = c("CD3D","CD79A"), split.by = "orig.ident")

VlnPlot2(pbmc, features = c("CD3D","CD79A"), violin = F)

}
