% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{write.py.fun}
\alias{write.py.fun}
\title{FUNCTION_TITLE}
\usage{
write.py.fun(fun, next.line = T, ...)
}
\arguments{
\item{fun}{PARAM_DESCRIPTION}

\item{next.line}{PARAM_DESCRIPTION, Default: T}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
