% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Python.R
\name{activate_python}
\alias{activate_python}
\title{Activate Python Environment for SeuratExtend}
\usage{
activate_python(conda_env = "seuratextend", verbose = TRUE, packages = c("os"))
}
\arguments{
\item{conda_env}{Character string specifying the conda environment name. Default is "seuratextend".}

\item{verbose}{Logical indicating whether to print status messages. Default is TRUE.}

\item{packages}{Character vector of Python packages to import.}
}
\value{
Invisible TRUE if successful
}
\description{
This function activates a Python environment (default: "seuratextend")
}
\examples{
\dontrun{
activate_python()
activate_python(conda_env = "my_custom_env", verbose = TRUE)
}
}
