% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistrBar.R
\name{BarOfCluster}
\alias{BarOfCluster}
\title{Cluster proportion bar plot}
\usage{
BarOfCluster(meta, cluster, origin, method = "value", do.heatmap = T)
}
\arguments{
\item{meta}{meta data data.frame}

\item{cluster}{variable name of cluster}

\item{origin}{variable name of sample}

\item{method}{'percent'or 'value'}

\item{do.heatmap}{Generate plot (TRUE) or matrix (FALSE), Default: T}
}
\value{
ggplot object or matrix
}
\description{
deprecated function of \code{\link[SeuratExtend:ClusterDistrBar]{ClusterDistrBar()}}
}
\examples{
meta <- seu@meta.data
origin <- meta$orig.ident
cluster <- meta$cluster
BarOfCluster(meta, "cluster", "orig.ident", method = "percent", do.heatmap = T)
}
\seealso{
\code{\link[SeuratExtend:ClusterDistrBar]{ClusterDistrBar()}}
}
