% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAUCRank.R
\name{BuildAUCRank}
\alias{BuildAUCRank}
\title{FUNCTION_TITLE}
\usage{
BuildAUCRank(seu, slot = "counts", assay = "RNA")
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{slot}{PARAM_DESCRIPTION, Default: 'counts'}

\item{assay}{PARAM_DESCRIPTION, Default: 'RNA'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
