% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUC_matrix.R
\name{calcAUC_matrix}
\alias{calcAUC_matrix}
\title{FUNCTION_TITLE}
\usage{
calcAUC_matrix(
  GenesetList,
  rankings,
  nCores = NULL,
  verbose = TRUE,
  n.items.part = 5e+05/ncol(rankings) * parallel::detectCores()
)
}
\arguments{
\item{GenesetList}{PARAM_DESCRIPTION}

\item{rankings}{PARAM_DESCRIPTION}

\item{nCores}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{n.items.part}{PARAM_DESCRIPTION, Default: 5e+05/ncol(rankings) * parallel::detectCores()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}
 \code{\link[AUCell]{AUCell_calcAUC}}
}
