% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcScoreGeneral.R
\name{CalcScoreGeneral}
\alias{CalcScoreGeneral}
\title{FUNCTION_TITLE}
\usage{
CalcScoreGeneral(matr, f, method = "zscore", exp.transform = F)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'zscore'}

\item{exp.transform}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
