% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcStats.R
\name{CalcScoreGeneral}
\alias{CalcScoreGeneral}
\alias{CalcScoreGeneral_v3}
\alias{ScoreAndOrder}
\title{CalcScoreGeneral, ScoreAndOrder}
\usage{
CalcScoreGeneral(
  matr,
  f,
  method = "zscore",
  exp.transform = F,
  t = F,
  order = NULL,
  n = Inf,
  p.threshold = 0.05
)

CalcScoreGeneral_v3(
  seu,
  features,
  group.by = NULL,
  cells = NULL,
  slot = "data",
  assay = NULL,
  method = "zscore",
  exp.transform = F,
  order = NULL,
  n = Inf,
  p.threshold = 0.05,
  priority = c("expr", "none")
)

ScoreAndOrder(
  matr,
  f,
  method = "zscore",
  exp.transform = F,
  t = F,
  order = NULL,
  n = Inf,
  p.threshold = 0.05
)
}
\description{
Alias of \code{\link[SeuratExtend:CalcStats]{CalcStats()}}
}
\seealso{
\code{\link[SeuratExtend:CalcStats]{CalcStats()}}
}
