% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{CellphoneDB_Plots}
\alias{CellphoneDB_Plots}
\title{FUNCTION_TITLE}
\usage{
CellphoneDB_Plots(
  seu = NULL,
  output = NULL,
  sender = NULL,
  receiver = NULL,
  group.by = NULL,
  top_n = 20,
  Dotplot = F
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION, Default: NULL}

\item{sender}{PARAM_DESCRIPTION, Default: NULL}

\item{receiver}{PARAM_DESCRIPTION, Default: NULL}

\item{group.by}{PARAM_DESCRIPTION, Default: NULL}

\item{top_n}{PARAM_DESCRIPTION, Default: 20}

\item{Dotplot}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
