% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistrBar.R
\name{ClusterDistrBar}
\alias{ClusterDistrBar}
\title{FUNCTION_TITLE}
\usage{
ClusterDistrBar(
  origin,
  cluster,
  rev = F,
  normalize = rev,
  percent = T,
  plot = T,
  flip = T,
  width = 0.9
)
}
\arguments{
\item{origin}{PARAM_DESCRIPTION}

\item{cluster}{PARAM_DESCRIPTION}

\item{rev}{PARAM_DESCRIPTION, Default: F}

\item{normalize}{PARAM_DESCRIPTION, Default: rev}

\item{percent}{PARAM_DESCRIPTION, Default: T}

\item{plot}{PARAM_DESCRIPTION, Default: T}

\item{flip}{PARAM_DESCRIPTION, Default: T}

\item{width}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc))

# absolute cell count
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc), percent = F)

# reverse x and y axis, normalized by sample size
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc), rev = T, normalize = T)

# reverse x and y axis, not normalized by sample size
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc), rev = T, normalize = F)

# vertical bar plot
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc), flip = F)

# export matrix
ClusterDistrBar(origin = pbmc$orig.ident, cluster = Idents(pbmc), plot = F)
}
