% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysisGO.R,
%   R/GeneSetAnalysisReactome.R
\name{FilterGOTerms}
\alias{FilterGOTerms}
\alias{FilterReactomeTerms}
\title{Filter GO/Reactome genesets}
\usage{
FilterGOTerms(
  term,
  spe = getOption("spe"),
  n.min = 1,
  n.max = Inf,
  only.end.terms = F,
  change.name = F,
  parent = NULL
)

FilterReactomeTerms(
  term,
  spe = getOption("spe"),
  n.min = 1,
  n.max = Inf,
  only.end.terms = F,
  change.name = F,
  parent = NULL
)
}
\arguments{
\item{term}{GO/Reactome IDs, can be character or matrix (ID in rownames)}

\item{spe}{Species (human or mouse), Default: getOption("spe")}

\item{n.min}{Min number of genes in the geneset, Default: 1}

\item{n.max}{Max number of genes in the geneset, Default: Inf}

\item{only.end.terms}{If only the end-level pathways/genesets are used for calculation, Default: F}

\item{change.name}{If to change names, Default: F}

\item{parent}{ID or name of parent (top-level) geneset in GO/Reactome database, Default: NULL}
}
\value{
character or matrix, depends on the input
}
\description{
Filter GO/Reactome genesets IDs
}
\examples{
options(max.print = 100, spe = "human")

# Filter GO terms
terms <- names(SeuratExtendData::GO_Data$human$GO2Gene)
terms
terms <- FilterGOTerms(terms, n.max = 1000, n.min = 10, only.end.terms = T, parent = "GO:0002376")
RenameGO(terms)

# Filter Reactome terms
terms <- names(SeuratExtendData::Reactome_Data$human$Path2Gene)
terms
terms <- FilterReactomeTerms(terms, n.max = 1000, n.min = 10, only.end.terms = T, parent = "R-HSA-168256")
RenameReactome(terms)
}
