% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysisGO.R
\name{FilterGOTerms}
\alias{FilterGOTerms}
\title{FUNCTION_TITLE}
\usage{
FilterGOTerms(
  term,
  spe = getOption("spe"),
  n.min = 1,
  n.max = Inf,
  only.end.terms = T,
  change.name = F,
  parent = NULL
)
}
\arguments{
\item{term}{PARAM_DESCRIPTION}

\item{spe}{PARAM_DESCRIPTION, Default: 'mouse'}

\item{n.min}{PARAM_DESCRIPTION, Default: 1}

\item{n.max}{PARAM_DESCRIPTION, Default: Inf}

\item{only.end.terms}{PARAM_DESCRIPTION, Default: T}

\item{change.name}{PARAM_DESCRIPTION, Default: F}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
