% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysis.R
\name{GeneSetAnalysis}
\alias{GeneSetAnalysis}
\title{FUNCTION_TITLE}
\usage{
GeneSetAnalysis(
  seu = NULL,
  genesets,
  title = "genesets",
  ratio = 0.4,
  n.min = 1,
  n.max = Inf,
  slot = "counts",
  assay = "RNA",
  nCores = getOption("nCores"),
  export_to_matrix = F
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION, Default: NULL}

\item{genesets}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION, Default: 'genesets'}

\item{ratio}{PARAM_DESCRIPTION, Default: 0.4}

\item{n.min}{PARAM_DESCRIPTION, Default: 1}

\item{n.max}{PARAM_DESCRIPTION, Default: Inf}

\item{slot}{PARAM_DESCRIPTION, Default: 'counts'}

\item{assay}{PARAM_DESCRIPTION, Default: 'RNA'}

\item{nCores}{PARAM_DESCRIPTION, Default: getOption("nCores")}

\item{export_to_matrix}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[parallel]{detectCores}}
}
