% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysis.R, R/GeneSetAnalysisGO.R,
%   R/GeneSetAnalysisReactome.R
\name{GeneSetAnalysis}
\alias{GeneSetAnalysis}
\alias{GeneSetAnalysisGO}
\alias{GeneSetAnalysisReactome}
\title{Gene Set Enrichment Analysis}
\usage{
GeneSetAnalysis(
  seu = NULL,
  genesets,
  title = "genesets",
  ratio = 0.4,
  n.min = 1,
  n.max = Inf,
  slot = "counts",
  assay = "RNA",
  nCores = 1,
  aucMaxRank = NULL,
  export_to_matrix = F,
  verbose = TRUE,
  n.items.part = NULL
)

GeneSetAnalysisGO(
  seu = NULL,
  dataset = NULL,
  parent = NULL,
  root = "BP",
  spe = getOption("spe"),
  ratio = 0.4,
  n.min = 1,
  n.max = Inf,
  only.end.terms = F,
  slot = "counts",
  assay = "RNA",
  nCores = 1,
  aucMaxRank = NULL,
  title = NULL,
  export_to_matrix = F,
  verbose = TRUE,
  n.items.part = NULL
)

GeneSetAnalysisReactome(
  seu = NULL,
  parent = "All",
  spe = getOption("spe"),
  ratio = 0.4,
  n.min = 1,
  n.max = Inf,
  only.end.terms = F,
  slot = "counts",
  assay = "RNA",
  nCores = 1,
  aucMaxRank = NULL,
  title = NULL,
  export_to_matrix = F,
  verbose = TRUE,
  n.items.part = NULL
)
}
\arguments{
\item{seu}{Seurat object}

\item{genesets}{List of gene-sets (or signatures) to test in the cells.
The gene-sets should be provided as character list.}

\item{title}{Name of slot where the data will be saved}

\item{ratio}{Minimum ratio of genes in the geneset detected in the datasets, Default: 0.4}

\item{n.min}{Min number of genes in the geneset, Default: 1}

\item{n.max}{Max number of genes in the geneset, Default: Inf}

\item{slot}{Slot to pull feature data for, Default: 'counts'}

\item{assay}{Name of assay to use, Default: 'RNA'}

\item{nCores}{Number of cores to use for computation. Default: 1}

\item{aucMaxRank}{Threshold to calculate the AUC}

\item{export_to_matrix}{If TRUE, then return a AUCell matrix instead of Seurat object}

\item{verbose}{Should the function show progress messages? Default: TRUE}

\item{n.items.part}{If the datasets/genesets are huge, split the genesets into n parts to
let it run with lower RAM}

\item{dataset}{(GO) Alias of 'parent', Default: NULL}

\item{parent}{ID or name of parent (top-level) geneset in GO/Reactome database,
so that only the subset of genesets will be calculated. Default: NULL}

\item{root}{(GO) Which root to use (BP - Biological process, MF - Molecular Function,
CC - Cellular Component), Default: 'BP'}

\item{spe}{Species (human or mouse), Default: getOption("spe")}

\item{only.end.terms}{If only the end-level pathways/genesets are used for calculation, Default: F}
}
\value{
Seurat object or Matrix
}
\description{
Calculate GSEA score of gene sets at single cell level, By 'AUCell' package:

Aibar et al. (2017) SCENIC: single-cell regulatory network inference and clustering.
Nature Methods. doi: 10.1038/nmeth.4463

Aibar. et al. (2016) AUCell: Analysis of 'gene set' activity in single-cell RNA-seq data.
R/Bioconductor package.

Can use either customized genesets, or pre-built GO or Reactome database
}
\details{
If return Seurat object, AUCell matrix is saved in seu@misc[["AUCell"]][[title]]
}
\examples{
options(spe = "human")

# Geneset enrichment analysis (GSEA) using customized genesets
pbmc <- GeneSetAnalysis(pbmc, genesets = hall50$human)
matr <- pbmc@misc$AUCell$genesets
Heatmap(CalcStats(matr, f = pbmc$cluster), lab_fill = "zscore")

# GSEA using GO database
pbmc <- GeneSetAnalysisGO(pbmc, parent = "immune_system_process")
 <- pbmc@misc$AUCell$GO$immune_system_process
matr <- RenameGO(matr)
Heatmap(CalcStats(matr, f = pbmc$cluster, order = "p", n = 5), lab_fill = "zscore")

# GSEA using Reactome database
pbmc <- GeneSetAnalysisReactome(pbmc, parent = "Immune System")
matr <- pbmc@misc$AUCell$Reactome$`Immune System`
matr <- RenameReactome(matr)
Heatmap(CalcStats(matr, f = pbmc$cluster, order = "p", n = 5), lab_fill = "zscore")

}
