% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{Heatmap}
\usage{
Heatmap(
  score,
  color_scheme = c(low = muted("blue"), mid = "white", high = muted("red")),
  border_color = NULL,
  lab_fill = "score",
  angle = 45,
  hjust = 1,
  vjust = 1,
  legend_position = "right",
  feature_text_subset = NULL,
  segment.width = c(1, 2.5, 1),
  segment.size = 0.2,
  text.spacing = 2,
  text.size = 2.5,
  hide_axis_line = TRUE,
  expand_limits_x = NULL,
  facet_col = NULL,
  facet_row = NULL,
  panel.spacing = unit(5, "pt"),
  strip.placement = "outside",
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{score}{A matrix for input, for instance, one generated using the CalcStats function.}

\item{color_scheme}{Color scheme for the heatmap. Default: c(low = muted("blue"), mid = "white", high = muted("red")).}

\item{border_color}{Color for the tile borders. Default: NULL.}

\item{lab_fill}{Label for the color. Default: 'score'.}

\item{angle}{Angle of the x-axis text. Passed to element_text(). Default: 45.}

\item{hjust}{Horizontal justification of x-axis text. Passed to element_text(). Default: 1.}

\item{vjust}{Vertical justification of x-axis text. Passed to element_text(). Default: 1.}

\item{legend_position}{Position of the figure legend. Default: 'right'.}

\item{feature_text_subset}{A subset of feature names to be shown. Useful when there are many features. Default: NULL.}

\item{segment.width}{Adjust the width ratio of the line connecting heatmap rows and feature text. Only effective when `feature_text_subset` is set. Default: c(1, 2.5, 1).}

\item{segment.size}{Thickness of the line connecting heatmap rows and feature text. Only effective when `feature_text_subset` is set. Default: 0.2.}

\item{text.spacing}{Spacing between feature texts. Only effective when `feature_text_subset` is set. Default: 2.}

\item{text.size}{Font size of the feature text. Only effective when `feature_text_subset` is set. Default: 2.5.}

\item{hide_axis_line}{Whether to hide the axis line or not. Default: TRUE.}

\item{expand_limits_x}{Sometimes the first name on the x-axis (e.g., cluster name) might be too long and surpass the left boundary of the plot. Expand the x-axis's left boundary to display the entire cluster name. Default: NULL.}

\item{facet_col}{Vector or factor to split the heatmap columns. Default: NULL.}

\item{facet_row}{Vector or factor to split the heatmap rows. Default: NULL.}

\item{panel.spacing}{Spacing between panels when rows or columns are split. Default: unit(5, "pt").}

\item{strip.placement}{Placement of panel names when rows or columns are split. Default: 'outside'.}

\item{ncol}{Number of columns when either `facet_col` or `facet_row` is not NULL, wrapping the heatmap. Default: NULL.}

\item{nrow}{Number of rows when either `facet_col` or `facet_row` is not NULL, wrapping the heatmap. Default: NULL.}

\item{...}{Additional parameters passed to the `theme` function of ggplot2.}
}
\value{
A ggplot object.
}
\description{
Generates a heatmap plot.
}
\details{
For more detailed usage, see the examples provided.
}
\examples{
# First, create a matrix using the CalcStats function.
genes <- VariableFeatures(pbmc)
toplot <- CalcStats(pbmc, features = genes, method = "zscore", order = "p", n = 5)

# Generate a basic heatmap.
Heatmap(toplot, lab_fill = "zscore")

# Modify the color theme to range from white to dark green.
Heatmap(toplot, lab_fill = "zscore", color_scheme = c("white", muted("green")))

# Use a color theme that transitions from dark blue to light yellow (centered at 0) to dark red.
Heatmap(toplot, lab_fill = "zscore", color_scheme = c(
  low = muted("blue"),
  mid = "lightyellow",
  high = muted("red"))
)

# Employ the viridis color theme, options include A, B, C, D, or E.
Heatmap(toplot, lab_fill = "zscore", color_scheme = "A")

# Expand the x-axis left boundary to ensure the full name of the first cluster is visible.
Heatmap(toplot, lab_fill = "zscore", expand_limits_x = -0.5)

# Construct a dense matrix with data for 500 genes.
toplot2 <- CalcStats(pbmc, features = genes[1:500], method = "zscore", order = "p")

# Display only a subset of gene names.
Heatmap(toplot2, lab_fill = "zscore", feature_text_subset = genes[1:20], expand_limits_x = c(-0.5, 11))

# Divide the heatmap into rows based on gene groups.
gene_groups <- sample(c("group1", "group2", "group3"), nrow(toplot2), replace = TRUE)
Heatmap(toplot2, lab_fill = "zscore", facet_row = gene_groups) +
  theme(axis.text.y = element_blank())

}
