% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{FUNCTION_TITLE}
\usage{
Heatmap(
  score,
  color_scheme = c(low = muted("blue"), mid = "white", high = muted("red")),
  border_color = "white",
  lab_fill = "score",
  angle = 45,
  hjust = 1,
  vjust = 1,
  legend_position = "right",
  feature_text_subset = NULL,
  nudge_x = ncol(score)/10,
  hide_axis_line = TRUE,
  expand_limits_x = NULL,
  facet_col = NULL,
  facet_row = NULL,
  panel.spacing = unit(5, "pt"),
  strip.placement = "outside"
)
}
\arguments{
\item{score}{PARAM_DESCRIPTION}

\item{color_scheme}{PARAM_DESCRIPTION, Default: c(low = muted("blue"), mid = "white", high = muted("red"))}

\item{border_color}{PARAM_DESCRIPTION, Default: 'white'}

\item{lab_fill}{PARAM_DESCRIPTION, Default: 'score'}

\item{angle}{PARAM_DESCRIPTION, Default: 45}

\item{hjust}{PARAM_DESCRIPTION, Default: 1}

\item{vjust}{PARAM_DESCRIPTION, Default: 1}

\item{legend_position}{PARAM_DESCRIPTION, Default: 'right'}

\item{feature_text_subset}{PARAM_DESCRIPTION, Default: NULL}

\item{nudge_x}{PARAM_DESCRIPTION, Default: ncol(score)/10}

\item{hide_axis_line}{PARAM_DESCRIPTION, Default: TRUE}

\item{expand_limits_x}{PARAM_DESCRIPTION, Default: NULL}

\item{facet_col}{PARAM_DESCRIPTION, Default: NULL}

\item{facet_row}{PARAM_DESCRIPTION, Default: NULL}

\item{panel.spacing}{PARAM_DESCRIPTION, Default: unit(5, "pt")}

\item{strip.placement}{PARAM_DESCRIPTION, Default: 'outside'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
