% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{Heatmap}
\alias{Heatmap}
\title{FUNCTION_TITLE}
\usage{
Heatmap(score, color_scheme = "2", lab_fill = "score", angle = 45, hjust = 1)
}
\arguments{
\item{score}{PARAM_DESCRIPTION}

\item{color_scheme}{PARAM_DESCRIPTION, Default: '2'}

\item{lab_fill}{PARAM_DESCRIPTION, Default: 'score'}

\item{angle}{PARAM_DESCRIPTION, Default: 45}

\item{hjust}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
