% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenic.R
\name{ImportPyscenicLoom}
\alias{ImportPyscenicLoom}
\title{FUNCTION_TITLE}
\usage{
ImportPyscenicLoom(loom.path, seu = NULL)
}
\arguments{
\item{loom.path}{PARAM_DESCRIPTION}

\item{seu}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
