% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nichenetr.R
\name{Nichenetr_plots}
\alias{Nichenetr_plots}
\title{FUNCTION_TITLE}
\usage{
Nichenetr_plots(
  seu.sender,
  seu.receiver,
  ident.snd = NULL,
  ident.rcv = NULL,
  pct.snd = 0.25,
  pct.rcv = 0.1,
  feature.logfc.threshold.rcv = 0.5,
  n_best_ligands = 30,
  n_ligand_target_links = 200,
  local.path.nichenetr.db = "Nichenetr",
  save.db = T,
  spe = getOption("spe")
)
}
\arguments{
\item{seu.sender}{PARAM_DESCRIPTION}

\item{seu.receiver}{PARAM_DESCRIPTION}

\item{ident.snd}{PARAM_DESCRIPTION, Default: NULL}

\item{ident.rcv}{PARAM_DESCRIPTION, Default: NULL}

\item{pct.snd}{PARAM_DESCRIPTION, Default: 0.25}

\item{pct.rcv}{PARAM_DESCRIPTION, Default: 0.1}

\item{feature.logfc.threshold.rcv}{PARAM_DESCRIPTION, Default: 0.5}

\item{n_best_ligands}{PARAM_DESCRIPTION, Default: 30}

\item{n_ligand_target_links}{PARAM_DESCRIPTION, Default: 200}

\item{local.path.nichenetr.db}{PARAM_DESCRIPTION, Default: 'Nichenetr'}

\item{save.db}{PARAM_DESCRIPTION, Default: T}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{remote-reexports}}
 \code{\link[dplyr]{desc}},\code{\link[dplyr]{select}}
 \code{\link[magrittr]{extract}}
 \code{\link[cowplot]{plot_grid}}
 \code{\link[ggpubr]{as_ggplot}},\code{\link[ggpubr]{get_legend}}
}
