% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameGO.R
\name{RenameGO}
\alias{RenameGO}
\title{FUNCTION_TITLE}
\usage{
RenameGO(term)
}
\arguments{
\item{term}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
