% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysisGO.R,
%   R/GeneSetAnalysisReactome.R
\name{RenameGO}
\alias{RenameGO}
\alias{RenameReactome}
\title{Add full name to GO/Reactome pathway ID}
\usage{
RenameGO(term, add_id = T, add_n_gene = T, spe = getOption("spe"))

RenameReactome(term, add_id = T, add_n_gene = T, spe = getOption("spe"))
}
\arguments{
\item{term}{GO/Reactome IDs, can be character or matrix (ID in rownames)}

\item{add_id}{If show IDs, Default: T}

\item{add_n_gene}{If show how many genes in each geneset, Default: T}

\item{spe}{Species (human or mouse), Default: getOption("spe")}
}
\value{
character or matrix, depends on the input
}
\description{
Rename GO/Reactome ID to full geneset/pathway name
}
\details{
Also see examples in \code{\link[SeuratExtend:GeneSetAnalysis]{GeneSetAnalysis()}}
}
\examples{
RenameGO(c("GO:0002376","GO:0050896"), spe = "human")

RenameReactome(c("R-HSA-109582","R-HSA-112316"), spe = "human")
}
