% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameGO.R
\name{RenameGO}
\alias{RenameGO}
\title{FUNCTION_TITLE}
\usage{
RenameGO(term, add_id = T, add_n_gene = T, spe = getOption("spe"))
}
\arguments{
\item{term}{PARAM_DESCRIPTION}

\item{add_id}{PARAM_DESCRIPTION, Default: T}

\item{add_n_gene}{PARAM_DESCRIPTION, Default: T}

\item{spe}{PARAM_DESCRIPTION, Default: getOption("spe")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
