% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellphoneDB.R
\name{RunCellphoneDB}
\alias{RunCellphoneDB}
\title{FUNCTION_TITLE}
\usage{
RunCellphoneDB(
  seu,
  group.by,
  cpdb.dir = NULL,
  method = c("analysis", "statistical_analysis"),
  database = "cellphonedb_mouse_cytokines",
  export.to.Seurat = T,
  NormalizeData = T
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION}

\item{cpdb.dir}{PARAM_DESCRIPTION, Default: NULL}

\item{method}{PARAM_DESCRIPTION, Default: c("analysis", "statistical_analysis")}

\item{database}{PARAM_DESCRIPTION, Default: 'cellphonedb_mouse_cytokines'}

\item{export.to.Seurat}{PARAM_DESCRIPTION, Default: T}

\item{NormalizeData}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
