% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palantir.R
\name{RunPalantirDiffusionMap}
\alias{RunPalantirDiffusionMap}
\title{FUNCTION_TITLE}
\usage{
RunPalantirDiffusionMap(seu, reduction = "pca", n_components = 20)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{reduction}{PARAM_DESCRIPTION, Default: 'pca'}

\item{n_components}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
