% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVelo.R
\name{scVelo.SeuratToLoom}
\alias{scVelo.SeuratToLoom}
\title{FUNCTION_TITLE}
\usage{
scVelo.SeuratToLoom(
  seu,
  export.name,
  velocyto.loom.path,
  velocyto.loom.filenames,
  fxn.convert.loomtoseurat.cellname = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{export.name}{PARAM_DESCRIPTION}

\item{velocyto.loom.path}{PARAM_DESCRIPTION}

\item{velocyto.loom.filenames}{PARAM_DESCRIPTION}

\item{fxn.convert.loomtoseurat.cellname}{PARAM_DESCRIPTION, Default: NULL}

\item{prefix}{PARAM_DESCRIPTION, Default: NULL}

\item{postfix}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
