% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVelo.R
\name{scVelo.SeuratToLoom}
\alias{scVelo.SeuratToLoom}
\title{Export loom for scVelo}
\usage{
scVelo.SeuratToLoom(
  seu,
  export.name,
  velocyto.loom.path = NULL,
  velocyto.loom.filenames,
  fxn.convert.loomtoseurat.cellname = NULL,
  cell.id.match.table = NULL,
  prefix = NULL,
  postfix = NULL
)
}
\arguments{
\item{seu}{Seurat object}

\item{export.name}{Name of loom file}

\item{velocyto.loom.path}{Directory to velocyto output (loom file), Default: getwd()}

\item{velocyto.loom.filenames}{File names of velocyto output (loom file)}

\item{fxn.convert.loomtoseurat.cellname}{(Experimental) function to convert velocyto
cell names to Seurat cell names}

\item{prefix}{Seurat cell id prefix to barcode (usually sample name)}

\item{postfix}{Seurat cell id postfix to barcode (usually '-1')}
}
\value{
a loom file
}
\description{
Export Seurat object to loom for scVelo run, and merge
velocyto output (spliced/unspliced matrix) into the the loom
}
\details{
See above
}
