% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin}
\alias{StackedViolin}
\title{FUNCTION_TITLE}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  ncol = 1,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  pt = F,
  pt.style = c("quasirandom", "jitter"),
  pt.size = 1.5,
  pt.alpha = 0.35
)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{f2}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: 1}

\item{lab_fill}{PARAM_DESCRIPTION, Default: 'group'}

\item{scales}{PARAM_DESCRIPTION, Default: 'free_y'}

\item{violin}{PARAM_DESCRIPTION, Default: T}

\item{box}{PARAM_DESCRIPTION, Default: T}

\item{width}{PARAM_DESCRIPTION, Default: 0.9}

\item{pt}{PARAM_DESCRIPTION, Default: F}

\item{pt.style}{PARAM_DESCRIPTION, Default: c("quasirandom", "jitter")}

\item{pt.size}{PARAM_DESCRIPTION, Default: 1.5}

\item{pt.alpha}{PARAM_DESCRIPTION, Default: 0.35}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
