% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vlnplots.R
\name{StackedViolin}
\alias{StackedViolin}
\alias{StackedViolin_v3}
\title{StackedViolin}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  t = F,
  ncol = 1,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  pt = T,
  hide.outlier = F,
  pt.style = c("jitter", "quasirandom"),
  pt.size = 1,
  pt.alpha = 1,
  strip.position = "top"
)

StackedViolin_v3(
  seu,
  features,
  group.by = NULL,
  split.by = NULL,
  cell = NULL,
  slot = "data",
  assay = NULL,
  priority = c("expr", "none"),
  ...
)
}
\description{
Alias of \code{\link[SeuratExtend:VlnPlot2]{VlnPlot2()}}
}
\seealso{
\code{\link[SeuratExtend:VlnPlot2]{VlnPlot2()}}
}
