% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin}
\alias{StackedViolin}
\title{FUNCTION_TITLE}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  ncol = 1,
  lab_fill = "group"
)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{f2}{PARAM_DESCRIPTION, Default: NULL}

\item{features}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: 1}

\item{lab_fill}{PARAM_DESCRIPTION, Default: 'group'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reshape2]{melt}}
}
