% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin}
\alias{StackedViolin}
\alias{StackedViolin_v3}
\title{StackedViolin}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  ncol = 1,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  pt = T,
  pt.style = c("jitter", "quasirandom"),
  pt.size = 1,
  pt.alpha = 0.35
)

StackedViolin_v3(
  seu,
  features,
  group.by = NULL,
  split.by = NULL,
  cell = NULL,
  ...
)
}
\description{
Alias of VlnPlot2
}
\seealso{
\code{\link[SeuratExtend:VlnPlot2]{VlnPlot2()}}
}
