% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vlnplots.R
\name{StackedViolin}
\alias{StackedViolin}
\alias{StackedViolin_v3}
\title{StackedViolin}
\usage{
StackedViolin(
  matr,
  f,
  f2 = NULL,
  features = NULL,
  t = F,
  cols = "pro_default",
  ncol = NULL,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  show.mean = FALSE,
  mean_colors = c("red", "blue"),
  pt = T,
  hide.outlier = F,
  pt.style = c("jitter", "quasirandom"),
  pt.size = 0.2,
  pt.alpha = 1,
  strip.position = "top",
  stat.method = c("none", "wilcox.test", "t.test"),
  stats.method = NULL,
  p.adjust.method = "holm",
  label = c("p.signif", "p", "p.adj", "p.format"),
  comparisons = NULL,
  hide.ns = TRUE,
  step.increase = 0.12,
  tip.length = 0.03,
  angle = NULL,
  hjust = NULL,
  vjust = NULL,
  ...
)

StackedViolin_v3(
  seu,
  features,
  group.by = NULL,
  split.by = NULL,
  cells = NULL,
  slot = "data",
  assay = NULL,
  priority = c("expr", "none"),
  cols = "auto",
  load.cols = TRUE,
  ncol = NULL,
  lab_fill = "group",
  scales = "free_y",
  violin = T,
  box = T,
  width = 0.9,
  show.mean = FALSE,
  mean_colors = c("red", "blue"),
  pt = T,
  hide.outlier = F,
  pt.style = c("jitter", "quasirandom"),
  pt.size = 0.2,
  pt.alpha = 1,
  strip.position = "top",
  stat.method = c("none", "wilcox.test", "t.test"),
  stats.method = NULL,
  p.adjust.method = "holm",
  label = c("p.signif", "p", "p.adj", "p.format"),
  comparisons = NULL,
  hide.ns = TRUE,
  step.increase = 0.12,
  tip.length = 0.03,
  angle = NULL,
  hjust = NULL,
  vjust = NULL,
  ...
)
}
\description{
Alias of \code{\link[SeuratExtend:VlnPlot2]{VlnPlot2()}}
}
\seealso{
\code{\link[SeuratExtend:VlnPlot2]{VlnPlot2()}}
}
