% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedViolin.R
\name{StackedViolin_v3}
\alias{StackedViolin_v3}
\title{FUNCTION_TITLE}
\usage{
StackedViolin_v3(
  seu,
  features,
  group.by = "seurat_clusters",
  split.by = NULL,
  cell = NULL,
  ncol = 1,
  scales = "free_y",
  type = c("violin", "boxplot"),
  outlier.size = 1
)
}
\arguments{
\item{seu}{PARAM_DESCRIPTION}

\item{features}{PARAM_DESCRIPTION}

\item{group.by}{PARAM_DESCRIPTION, Default: 'seurat_clusters'}

\item{split.by}{PARAM_DESCRIPTION, Default: NULL}

\item{cell}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: 1}

\item{scales}{PARAM_DESCRIPTION, Default: 'free_y'}

\item{type}{PARAM_DESCRIPTION, Default: c("violin", "boxplot")}

\item{outlier.size}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
