% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesymbolConversion.R
\name{UniprotToGenesymbol}
\alias{UniprotToGenesymbol}
\title{Convert UniProt IDs to Gene Symbols}
\usage{
UniprotToGenesymbol(Uniprot, spe = getOption("spe"))
}
\arguments{
\item{Uniprot}{A vector of UniProt IDs that are to be converted to gene symbols.}

\item{spe}{The species for which the conversion should be performed. The default is set to the user's global option (`getOption("spe")`), but can be explicitly specified as "human" or "mouse".}
}
\value{
Returns a vector of gene symbols corresponding to the input UniProt IDs.
}
\description{
Converts UniProt IDs to corresponding gene symbols for specified species using local databases. This function supports integration between proteomic and genomic data, enhancing the utility of proteomic identifiers in genomic analyses.
}
\details{
`UniprotToGenesymbol` utilizes a locally stored version of the UniProt database to convert UniProt IDs to gene symbols. This approach ensures fast and reliable conversions without the need for online database access. The local database includes comprehensive mappings from UniProt IDs to gene symbols for both human and mouse, facilitating seamless integration of proteomic and genomic data.
}
\examples{
# Converting UniProt IDs to human gene symbols
UniprotToGenesymbol(c("Q8NF67", "Q9NPB9"), spe = "human")

# Converting UniProt IDs to mouse gene symbols
UniprotToGenesymbol(c("Q9R1C8", "Q9QY84"), spe = "mouse")
}
