% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterfallPlot.R
\name{WaterfallPlot}
\alias{WaterfallPlot}
\title{FUNCTION_TITLE}
\usage{
WaterfallPlot(
  matr,
  f,
  ident.1 = NULL,
  ident.2 = NULL,
  exp.transform = F,
  length = "tscore",
  color = "tscore",
  len.threshold = 0,
  col.threshold = 0,
  flip = T,
  y.label = length,
  angle = NULL,
  hjust = NULL,
  vjust = NULL
)
}
\arguments{
\item{matr}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{ident.1}{PARAM_DESCRIPTION, Default: NULL}

\item{ident.2}{PARAM_DESCRIPTION, Default: NULL}

\item{exp.transform}{PARAM_DESCRIPTION, Default: F}

\item{length}{PARAM_DESCRIPTION, Default: 'tscore'}

\item{color}{PARAM_DESCRIPTION, Default: 'tscore'}

\item{len.threshold}{PARAM_DESCRIPTION, Default: 0}

\item{col.threshold}{PARAM_DESCRIPTION, Default: 0}

\item{flip}{PARAM_DESCRIPTION, Default: T}

\item{y.label}{PARAM_DESCRIPTION, Default: length}

\item{angle}{PARAM_DESCRIPTION, Default: NULL}

\item{hjust}{PARAM_DESCRIPTION, Default: NULL}

\item{vjust}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
