% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterfallPlot.R
\name{WaterfallPlot_v3}
\alias{WaterfallPlot_v3}
\title{WaterfallPlot_v3}
\usage{
WaterfallPlot_v3(
  seu,
  features,
  group.by = NULL,
  cell = NULL,
  slot = "data",
  assay = NULL,
  priority = c("expr", "none"),
  ident.1 = NULL,
  ident.2 = NULL,
  exp.transform = (length == "logFC"),
  order = TRUE,
  length = "logFC",
  color = "p",
  len.threshold = 0,
  col.threshold = 0,
  color_theme = "BuRd",
  center_color = NULL,
  top.n = NULL,
  flip = FALSE,
  y.label = NULL,
  angle = NULL,
  hjust = NULL,
  vjust = NULL,
  title = NULL
)
}
\description{
Alias of \code{\link[SeuratExtend:WaterfallPlot]{WaterfallPlot()}}, Seurat version
}
\seealso{
\code{\link[SeuratExtend:WaterfallPlot]{WaterfallPlot()}}
}
